[IMP/REM] Renamed module onestein_project_scrum_agile to project_scrum_agile_custom
