+++ custom/project_scrum_agile_extended/__manifest__.py
@@ -1,15 +1,11 @@
 # License LGPL-3.0 or later (https://www.gnu.org/licenses/lgpl.html).
 {
     # Module information
-    "name": "Onestein Project Scrum Management",
+    "name": "Project Scrum Management Extended",
     "version": "15.0.1.0.0",
     "license": "LGPL-3",
     "category": "Project Scrum Management",
     "summary": """
-        This application is used to manage total planned hours, progress,
-        weightage of sprint, sprint release, product backlog.
-        """,
-    "description": """
         This application is used to manage total planned hours, progress,
         weightage of sprint, sprint release, product backlog.
         This application used to manage logs in of sprint release from sprint.
@@ -28,6 +24,7 @@
         "views/project_scrum_view.xml",
         "views/account_analytic_line_view.xml",
         "views/project_view.xml",
+        "views/project_task_view.xml",
     ],
     # Technical
     "installable": True,
