+++ custom/project_scrum_agile_extended/models/__init__.py
@@ -3,3 +3,6 @@
 from . import project_scrum_release
 from . import project_scrum
 from . import project
+from . import project_task
+from . import project_scrum_meeting
+from . import project_scrum_product_backlog
