+++ custom/project_scrum_agile_extended/views/project_scrum_view.xml
@@ -42,19 +42,4 @@
         </field>
     </record>
 
-    <record id="view_task_timesheet_form2" model="ir.ui.view">
-        <field name="name">project.task.scrum.timesheet.form</field>
-        <field name="model">project.task</field>
-        <field name="inherit_id" ref="hr_timesheet.view_task_form2_inherited" />
-        <field name="arch" type="xml">
-            <xpath expr="//field[@name='user_id']" position="attributes">
-                <attribute name="invisible">0</attribute>
-            </xpath>
-            <field name="planned_hours" position="after">
-                <field name="estimate_adjustment" />
-                <field name="weightage" />
-            </field>
-        </field>
-    </record>
-
 </odoo>
