+++ custom/project_scrum_agile_extended/views/project_task_view.xml
@@ -1,0 +1,29 @@
+<odoo>
+
+    <record id="inherit_view_task_form2" model="ir.ui.view">
+        <field name="name">inherit.view.task.form2</field>
+        <field name="model">project.task</field>
+        <field name="inherit_id" ref="project.view_task_form2" />
+        <field name="arch" type="xml">
+            <field name="date_deadline" position="before">
+                <field name="schedule_date" required="1" />
+            </field>
+        </field>
+    </record>
+
+    <record id="view_task_timesheet_form2" model="ir.ui.view">
+        <field name="name">project.task.scrum.timesheet.form</field>
+        <field name="model">project.task</field>
+        <field name="inherit_id" ref="hr_timesheet.view_task_form2_inherited" />
+        <field name="arch" type="xml">
+            <xpath expr="//field[@name='user_id']" position="attributes">
+                <attribute name="invisible">0</attribute>
+            </xpath>
+            <field name="planned_hours" position="after">
+                <field name="estimate_adjustment" />
+                <field name="weightage" />
+            </field>
+        </field>
+    </record>
+
+</odoo>
